refactor(pipeline): make pipeline.ts a real static fixture file
Move the dynamic pipeline.ts generation (getPipelineFixtureContent()) to a
real file on disk at test-projects/pipeline/amplify/pipeline.ts. The stack
name uniqueness is now handled via PIPELINE_STACK_NAME environment variable
(with fallback 'amplify-pipeline' for local development).

The test project directory now looks exactly like what a real user has:
- amplify/backend.ts
- amplify/auth/resource.ts
- amplify/data/resource.ts
- amplify/hosting.ts
- amplify/pipeline.ts  <-- now a real file, reads stack name from env
- public/index.html
- package.json

The test setup passes PIPELINE_STACK_NAME via env when calling ampx deploy.
